Try to structure the regress scripts so they will work both with CMake and autotools
svn:revision:41906
svn:branch:trunk
svn:account:starseeker

+++ regress/gqa.sh
@@ -42,12 +42,17 @@ export PATH || (echo "This isn't sh."; sh $0 $*; kill $$)
 # PATH_TO_THIS, and THIS.
 . $1/regress/library.sh
 
-MGED="`ensearch mged/mged`"
+MGED="`ensearch mged`"
 if test ! -f "$MGED" ; then
     echo "Unable to find mged, aborting"
     exit 1
 fi
 
+GQABIN="`ensearch g_qa`"
+if test ! -f "$GQABIN" ; then
+    echo "Unable to find g_qa, aborting"
+    exit 1
+fi
 
 rm -f gqa.g density_table.txt gqa.log gqa_mged.log gqa.mged
 
@@ -121,7 +126,7 @@ EOF
 # open_box.r = 1000-576 = 424 m^3
 
 
-GQA="../src/gtools/g_qa -u m,m^3,kg -g 250mm-50mm -p"
+GQA="$GQABIN -u m,m^3,kg -g 250mm-50mm -p"
 export GQA
 
 STATUS=0
